rename transactions -> user_commands
