+++ src/user_commands/table_trait.rs
